Created a small utility which walks through all instructions in an Amiga executable using InstructionLengthDecoder
The program will fetch one new instruction word at a time from the file, and attempt to perform length decode on what has currently been seen. If the length decodes, it prints meta information and advances to the next instruction. If the length doesn't decode, it tries again with an extra instruction word's worth of look-ahead. If the decoder is unable to decode even with 16 words in view, it gives up.

This program is intended as a sanity check for the length decoder; give it a known assembly program, and see if it produces the same instruction sequence?