Partial implementation of OP -> uOP decomposer
The implementation so far only handles ADD <ea>,Dn.
It assigns execution resources for all register arguments.
It fills in 16/32-bit immediate values into the uOP's two extension words.
It does not handle operation size properly yet.
It does not handle memory access satisfactorily either.
It handles 68020+ addressing modes properly, injecting extra LOAD/LEA uOPs when necessary.

The OpWordDecoder and the pairability tester have been removed from the list of projects to build, as the former is obsolete and the latter should be rewritten to use the output of the uOP decomposer instead.