Introduced some detailed test cases for Op->uOP decoding. AguOutput/AguResult handling is still iffy.
The Op->uOP decoder now has testcases which verify its entire output. Test cases have so far only been built for the ADD <ea>,Dn instruction. Test cases are for all 68000 and 68020 addressing modes.

It is unclear exactly how the AGU's outputs should be handled. As things are right now, the AGU is able to output two values but there is no ExecutionResource that refers to each of these. It looks likely that the simplest way of making the pipeline forwarding work properly later on is to have a separate ExecutionResource for each of these.