Removed OpWordDecoder and obsolete parts of ExecutionResource & uOP
The OpWordDecoder was based on a flawed concept; it is not possible to fully identify an instruction's resource usage based on the first OpWord. Complex addressing modes require inspection of additional words in the instruction.

Also, it is not enough to look at the first OpWord to identify whether or not two instructions will pair; it is actually more straightforward to first break down the instruction into uOPs and do the analysis on the uOPs.
The pairability test remains inactive for the time being. Later on it should be rewritten to take in two uOPs and make a decision based on those.