Added decoding support for ADDQ and 3-bit immediate handling
The 3-bit immediate values are stored directly inside the UOp and not in the extensionWords. This allows instructions like ADDQ.L #1,$1234(a0) to execute in a single cycle. It's possible that the real hardware handles this in a more sneaky manner - perhaps packaging the immediate value in the second extensionWord, as that is not used by any single-UOp ADDQ/SUBQ variants - or perhaps the real hardware passes the original OpWord along all the time. There are no tests we can perform to get the final answer. Therefor, introducing an extra 3-bit immediate field into the UOp is probably the best way to go. It clutters all uses of UOps and all the testcases... but until we know for sure that the second extensionWord really is available for all 3-bit operands, it's better to stick with the current approach.