De-duplicated OpWord decode information tables
The OpWord decode tables used to be stored separately in the instruction length decoder and Op->UOp decoder. They remained largely identical though... and it was turning out that a single table, with some extra Op->UOp information, would be enough to serve the needs for both decoding types.
A shared table would give better test coverage as well. So, better merge the tables.