Added Op->UOp decoding of MOVE/MOVEA instructions; not 100% correct yet though
MOVE.L to register, all MOVEA, and all MOVE to memory, do not need to read the IeeB operand (nor trigger a mem read for the destination). The current instruction decoder will generate reads for all destination operands. This can be seen from the test cases that are currently failing.

Also, MOVEA is not identified as a distinct IEE-ALU operation yet. It probably should - otherwise handling MOVEA.W properly will be difficult.