Added support for flags generation to IEE-ALU for Add/Sub operation
IEE-ALU ops which generate flag values do so by generating an AND-mask and an OR-mask. The reason for  doing this rather than rewriting a flags register directly is that later on, when running pOEP & sOEP in parallel, the AND-mask/OR-mask style makes it possible to accumulate two flag-affecting results by performing each of the prescribed AND/OR operations in sequence.

There are now pretty good testcases for Add/AddA/Sub/SubA for the IEE-ALU.