TestSimpleOEP is now a standardized test case model for SimpleOEP
TestSimpleOEP used to be a program which performed a number of UOps and printed the result after each one. It has now been converted into a series of testcases.

In the future, there should also be a program that takes in an instruction stream and generates a detailed trace of the activity that occurs when running the stream.