Implemented Op->UOp decoding of BTST/BLCR/BCHG/BSET in a somewhat clumsy way
The decoding itself is correct. What's special about the single-bit ops is that, for example, BSET #14,D3 has a byte-size source operand but a longword-size IEE operation. The Op->UOp decoding trees don't support source operands with a different size yet, so that has been hacked by adding a unique DecodeOperand & EAEncoding & EAMode type for immediates that are explicitly byte-size.

(A similar problem exists for xxxA instructions.)

Perhaps it would be better to specify decoding rules for all operands and IEE operation separately?