OpWord is part of UOp; imm3bit & imm8bit operands are taken directly from OpWord
According to the Apr 1995 article in IEEE Micro, the processor will stuff "operation word, 32-bits of extension words, and 32-bits of early-decode information" for each register into the FIFO instruction buffer. The simulation has been changed to behave in the same way.

The OpWord is set to 0x0000 for micro-ops that are generated on-the-fly. Perhaps it would make sense to assign unique OpWords to these - re-using illegal opcodes for them?