Added Op->UOp decoding logic for BSR, BRA, Bcc, JSR, JMP in a clunky way
The OpWordDecodeInformation table has been split into two; one for general instructions, another for branches.
All the branches will emit real UOps that write to PC. It's not a great implementation but it should work.
In the future, it would probably make sense to take the output of the Op->UOp decoder and rewrite any operations that include branches to handle PC smarter, and work with the branch cache. The current implementation is just intended to make it possible to build an actually-working CPU with flow control.